[ADD & FIX] test_fat.c adicionado e modificado o tipo do indexFAT

+++ include/types.h
@@ -65,7 +65,7 @@
 	*/
 
     //Lista de clusters representados na FAT
-    WORD* indexFAT;
+    DWORD* indexFAT;
 
     struct descritor LDAA[MAX_FILES_OPEN];
